refactor: refactor and add readme
